# docs/api/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QualiDiagIQ API",
    "description": "\n        Motor de Diagnóstico Tributário Automatizado para a Reforma do Consumo.\n        \n        Parte do ecossistema Tributiq. Fornece análise de maturidade e conformidade\n        com base na EC 132/2023, LC 214/2025 e ABNT NBR 17301:2026.\n        ",
    "contact": {
      "name": "Equipe Tributiq",
      "email": "contato@tributiq.com.br"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "Conta na plataforma"
        ],
        "summary": "Login",
        "description": "Login contra a tabela `admins`; token inclui tenant para RLS futuro.",
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/cadastro": {
      "post": {
        "tags": [
          "Conta na plataforma"
        ],
        "summary": "Cadastrar conta na plataforma (nome, e-mail, senha)",
        "description": "Cria registro em `admins` com tenant dedicado e `perfil_conta` gratuito; devolve o mesmo contrato de `POST /auth/login` (JWT). Requer PostgreSQL acessível como em login (`DATABASE_URL`) ou Supabase configurado. Desabilite em produção com `QDI_CADASTRO_CONSULTOR_B2B_HABILITADO=false`.",
        "operationId": "cadastro_consultor_b2b_auth_cadastro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CadastroConsultorB2BRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/verificar-email/solicitar": {
      "post": {
        "tags": [
          "Conta na plataforma"
        ],
        "summary": "Solicitar código por e-mail",
        "description": "Envia OTP numérico para o endereço informado (SMTP configurável). Rate limit por e-mail entre reenvios. MVP: armazenamento em memória no processo da API.",
        "operationId": "solicitar_verificacao_email_auth_verificar_email_solicitar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitarVerificacaoEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolicitarVerificacaoEmailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/auth/verificar-email/confirmar": {
      "post": {
        "tags": [
          "Conta na plataforma"
        ],
        "summary": "Confirmar código do e-mail",
        "operationId": "confirmar_verificacao_email_auth_verificar_email_confirmar_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarVerificacaoEmailRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfirmarVerificacaoEmailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CadastroConsultorB2BRequest": {
        "properties": {
          "nome": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Nome",
            "description": "Nome exibido no painel."
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 8,
            "title": "Password",
            "description": "Senha com pelo menos 8 caracteres (bcrypt no servidor)."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "nome",
          "email",
          "password"
        ],
        "title": "CadastroConsultorB2BRequest",
        "description": "Cadastro mínimo para acesso ao painel (conta na plataforma)."
      },
...
      "ConfirmarVerificacaoEmailRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "codigo": {
            "type": "string",
            "maxLength": 8,
            "minLength": 4,
            "title": "Codigo",
            "description": "Código recebido por e-mail"
          }
        },
        "type": "object",
        "required": [
          "email",
          "codigo"
        ],
        "title": "ConfirmarVerificacaoEmailRequest"
      },
      "ConfirmarVerificacaoEmailResponse": {
        "properties": {
          "verificado": {
            "type": "boolean",
            "title": "Verificado"
          }
        },
        "type": "object",
        "required": [
          "verificado"
        ],
        "title": "ConfirmarVerificacaoEmailResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3,
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 256,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest"
      },
      "LoginResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "nome": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nome"
          },
          "perfil_conta": {
            "type": "string",
            "title": "Perfil Conta",
            "default": "gratuito"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type",
          "nome"
        ],
        "title": "LoginResponse"
      },
...
      "SolicitarVerificacaoEmailRequest": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          }
        },
        "type": "object",
        "required": [
          "email"
        ],
        "title": "SolicitarVerificacaoEmailRequest",
        "description": "Wizard passo 1 — confirmação de posse do inbox (LGPD)."
      },
      "SolicitarVerificacaoEmailResponse": {
        "properties": {
          "mensagem": {
            "type": "string",
            "title": "Mensagem"
          }
        },
        "type": "object",
        "required": [
          "mensagem"
        ],
        "title": "SolicitarVerificacaoEmailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}